backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
httpx==0.25.2
google-generativeai==0.3.2
python-multipart==0.0.6
pydantic==2.5.0
